rename build_function to build_functions
